tests: remove deprecated/skipped test_try
Move remaining SCM Helpers to test_scm_helpers
